pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repograph"
version = "0.1.0"
description = "Code intelligence knowledge graph for AI agents"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "cogdb>=3.0.7",
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.2; python_version < '3.13'",
    "tree-sitter-language-pack>=1.1.2; python_version >= '3.13'",
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "pathspec>=0.12.1",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "sse-starlette<3",
    "starlette<0.48",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
repograph = "repograph.api.app:main"
repograph-mcp = "repograph.mcp_server.server:main"
